SIX-frontend-005 : Change project structure
