+++ src/api/net-worth.json
