+++ src/css/style.css
@@ -36,7 +36,7 @@
 
 /* 🌌 默认暗色模式背景图 */
 body {
-  background-image: url('../stars.jpg');
+  background-image: url('../../public/images/stars.jpg');
   background-size: cover;
   background-position: center;
   background-repeat: no-repeat;
@@ -202,7 +202,7 @@ h2 {
     border-radius: 20px;
     padding: 15px;
     margin-bottom: 20px;
-    background-color: rgba(34, 73, 119, 0.4);
+    background-color: rgb(255, 255, 255);
     backdrop-filter: blur(3px);
   }
