+++ src/js/modules/theme.js
@@ -1,0 +1,5 @@
+export function initThemeToggle() {
+    document.getElementById("toggle-theme").addEventListener("click", () => {
+        document.body.classList.toggle("light-theme");
+    });
+}
